# Fft optimize