--- a/atcoder/convolution.hpp
+++ b/atcoder/convolution.hpp
@@ -14,95 +14,165 @@ namespace atcoder {
 
 namespace internal {
 
+template <class mint,
+          int g = internal::primitive_root<mint::mod()>,
+          internal::is_static_modint_t<mint>* = nullptr>
+struct fft_info {
+    static constexpr int rank2 = bsf_constexpr(mint::mod() - 1);
+    std::array<mint, rank2 + 1> root;   // root[i]^(2^i) == 1
+    std::array<mint, rank2 + 1> iroot;  // root[i] * iroot[i] == 1
+
+    std::array<mint, std::max(0, rank2 - 2 + 1)> rate2;
+    std::array<mint, std::max(0, rank2 - 2 + 1)> irate2;
+
+    std::array<mint, std::max(0, rank2 - 3 + 1)> rate3;
+    std::array<mint, std::max(0, rank2 - 3 + 1)> irate3;
+
+    fft_info() {
+        root[rank2] = mint(g).pow((mint::mod() - 1) >> rank2);
+        iroot[rank2] = root[rank2].inv();
+        for (int i = rank2 - 1; i >= 0; i--) {
+            root[i] = root[i + 1] * root[i + 1];
+            iroot[i] = iroot[i + 1] * iroot[i + 1];
+        }
+
+        {
+            mint prod = 1, iprod = 1;
+            for (int i = 0; i <= rank2 - 2; i++) {
+                rate2[i] = root[i + 2] * prod;
+                irate2[i] = iroot[i + 2] * iprod;
+                prod *= iroot[i + 2];
+                iprod *= root[i + 2];
+            }
+        }
+        {
+            mint prod = 1, iprod = 1;
+            for (int i = 0; i <= rank2 - 3; i++) {
+                rate3[i] = root[i + 3] * prod;
+                irate3[i] = iroot[i + 3] * iprod;
+                prod *= iroot[i + 3];
+                iprod *= root[i + 3];
+            }
+        }
+    }
+};
+
 template <class mint, internal::is_static_modint_t<mint>* = nullptr>
 void butterfly(std::vector<mint>& a) {
-    static constexpr int g = internal::primitive_root<mint::mod()>;
     int n = int(a.size());
     int h = internal::ceil_pow2(n);
 
-    static bool first = true;
-    static mint sum_e[30];  // sum_e[i] = ies[0] * ... * ies[i - 1] * es[i]
-    if (first) {
-        first = false;
-        mint es[30], ies[30];  // es[i]^(2^(2+i)) == 1
-        int cnt2 = bsf(mint::mod() - 1);
-        mint e = mint(g).pow((mint::mod() - 1) >> cnt2), ie = e.inv();
-        for (int i = cnt2; i >= 2; i--) {
-            // e^(2^i) == 1
-            es[i - 2] = e;
-            ies[i - 2] = ie;
-            e *= e;
-            ie *= ie;
-        }
-        mint now = 1;
-        for (int i = 0; i <= cnt2 - 2; i++) {
-            sum_e[i] = es[i] * now;
-            now *= ies[i];
-        }
-    }
-    for (int ph = 1; ph <= h; ph++) {
-        int w = 1 << (ph - 1), p = 1 << (h - ph);
-        mint now = 1;
-        for (int s = 0; s < w; s++) {
-            int offset = s << (h - ph + 1);
-            for (int i = 0; i < p; i++) {
-                auto l = a[i + offset];
-                auto r = a[i + offset + p] * now;
-                a[i + offset] = l + r;
-                a[i + offset + p] = l - r;
+    static const fft_info<mint> info;
+
+    int len = 0;  // a[i, i+(n>>len), i+2*(n>>len), ..] is transformed
+    while (len < h) {
+        if (h - len == 1) {
+            int p = 1 << (h - len - 1);
+            mint rot = 1;
+            for (int s = 0; s < (1 << len); s++) {
+                int offset = s << (h - len);
+                for (int i = 0; i < p; i++) {
+                    auto l = a[i + offset];
+                    auto r = a[i + offset + p] * rot;
+                    a[i + offset] = l + r;
+                    a[i + offset + p] = l - r;
+                }
+                rot *= info.rate2[bsf(~(unsigned int)(s))];
             }
-            now *= sum_e[bsf(~(unsigned int)(s))];
+            len++;
+        } else {
+            // 4-base
+            int p = 1 << (h - len - 2);
+            mint rot = 1, imag = info.root[2];
+            for (int s = 0; s < (1 << len); s++) {
+                mint rot2 = rot * rot;
+                mint rot3 = rot2 * rot;
+                int offset = s << (h - len);
+                for (int i = 0; i < p; i++) {
+                    auto mod2 = 1ULL * mint::mod() * mint::mod();
+                    auto a0 = 1ULL * a[i + offset].val();
+                    auto a1 = 1ULL * a[i + offset + p].val() * rot.val();
+                    auto a2 = 1ULL * a[i + offset + 2 * p].val() * rot2.val();
+                    auto a3 = 1ULL * a[i + offset + 3 * p].val() * rot3.val();
+                    auto a1na3imag =
+                        1ULL * mint(a1 + mod2 - a3).val() * imag.val();
+                    auto na2 = mod2 - a2;
+                    a[i + offset] = a0 + a2 + a1 + a3;
+                    a[i + offset + 1 * p] = a0 + a2 + (2 * mod2 - (a1 + a3));
+                    a[i + offset + 2 * p] = a0 + na2 + a1na3imag;
+                    a[i + offset + 3 * p] = a0 + na2 + (mod2 - a1na3imag);
+                }
+                rot *= info.rate3[bsf(~(unsigned int)(s))];
+            }
+            len += 2;
         }
     }
 }
 
 template <class mint, internal::is_static_modint_t<mint>* = nullptr>
 void butterfly_inv(std::vector<mint>& a) {
-    static constexpr int g = internal::primitive_root<mint::mod()>;
     int n = int(a.size());
     int h = internal::ceil_pow2(n);
 
-    static bool first = true;
-    static mint sum_ie[30];  // sum_ie[i] = es[0] * ... * es[i - 1] * ies[i]
-    if (first) {
-        first = false;
-        mint es[30], ies[30];  // es[i]^(2^(2+i)) == 1
-        int cnt2 = bsf(mint::mod() - 1);
-        mint e = mint(g).pow((mint::mod() - 1) >> cnt2), ie = e.inv();
-        for (int i = cnt2; i >= 2; i--) {
-            // e^(2^i) == 1
-            es[i - 2] = e;
-            ies[i - 2] = ie;
-            e *= e;
-            ie *= ie;
-        }
-        mint now = 1;
-        for (int i = 0; i <= cnt2 - 2; i++) {
-            sum_ie[i] = ies[i] * now;
-            now *= es[i];
-        }
-    }
+    static const fft_info<mint> info;
 
-    for (int ph = h; ph >= 1; ph--) {
-        int w = 1 << (ph - 1), p = 1 << (h - ph);
-        mint inow = 1;
-        for (int s = 0; s < w; s++) {
-            int offset = s << (h - ph + 1);
-            for (int i = 0; i < p; i++) {
-                auto l = a[i + offset];
-                auto r = a[i + offset + p];
-                a[i + offset] = l + r;
-                a[i + offset + p] =
-                    (unsigned long long)(mint::mod() + l.val() - r.val()) *
-                    inow.val();
+    int len = h;  // a[i, i+(n>>len), i+2*(n>>len), ..] is transformed
+    while (len) {
+        if (len == 1) {
+            int p = 1 << (h - len);
+            mint irot = 1;
+            for (int s = 0; s < (1 << (len - 1)); s++) {
+                int offset = s << (h - len + 1);
+                for (int i = 0; i < p; i++) {
+                    auto l = a[i + offset];
+                    auto r = a[i + offset + p];
+                    a[i + offset] = l + r;
+                    a[i + offset + p] =
+                        (unsigned long long)(mint::mod() + l.val() - r.val()) *
+                        irot.val();
+                    ;
+                }
+                irot *= info.irate2[bsf(~(unsigned int)(s))];
             }
-            inow *= sum_ie[bsf(~(unsigned int)(s))];
+            len--;
+        } else {
+            // 4-base
+            int p = 1 << (h - len);
+            mint irot = 1, iimag = info.iroot[2];
+            for (int s = 0; s < (1 << (len - 2)); s++) {
+                mint irot2 = irot * irot;
+                mint irot3 = irot2 * irot;
+                int offset = s << (h - len + 2);
+                for (int i = 0; i < p; i++) {
+                    auto a0 = 1ULL * a[i + offset + 0 * p].val();
+                    auto a1 = 1ULL * a[i + offset + 1 * p].val();
+                    auto a2 = 1ULL * a[i + offset + 2 * p].val();
+                    auto a3 = 1ULL * a[i + offset + 3 * p].val();
+
+                    auto a2na3iimag =
+                        1ULL *
+                        mint((mint::mod() + a2 - a3) * iimag.val()).val();
+
+                    a[i + offset] = a0 + a1 + a2 + a3;
+                    a[i + offset + 1 * p] =
+                        (a0 + (mint::mod() - a1) + a2na3iimag) * irot.val();
+                    a[i + offset + 2 * p] =
+                        (a0 + a1 + (mint::mod() - a2) + (mint::mod() - a3)) *
+                        irot2.val();
+                    a[i + offset + 3 * p] =
+                        (a0 + (mint::mod() - a1) + (mint::mod() - a2na3iimag)) *
+                        irot3.val();
+                }
+                irot *= info.irate3[bsf(~(unsigned int)(s))];
+            }
+            len -= 2;
         }
     }
 }
 
 template <class mint, internal::is_static_modint_t<mint>* = nullptr>
-std::vector<mint> convolution_naive(const std::vector<mint>& a, const std::vector<mint>& b) {
+std::vector<mint> convolution_naive(const std::vector<mint>& a,
+                                    const std::vector<mint>& b) {
     int n = int(a.size()), m = int(b.size());
     std::vector<mint> ans(n + m - 1);
     if (n < m) {
@@ -150,7 +220,8 @@ std::vector<mint> convolution(std::vector<mint>&& a, std::vector<mint>&& b) {
 }
 
 template <class mint, internal::is_static_modint_t<mint>* = nullptr>
-std::vector<mint> convolution(const std::vector<mint>& a, const std::vector<mint>& b) {
+std::vector<mint> convolution(const std::vector<mint>& a,
+                              const std::vector<mint>& b) {
     int n = int(a.size()), m = int(b.size());
     if (!n || !m) return {};
     if (std::min(n, m) <= 60) return convolution_naive(a, b);