--- a/test/unittest/convolution_test.cpp
+++ b/test/unittest/convolution_test.cpp
@@ -391,6 +391,19 @@ TEST(ConvolutionTest, Conv2) {
     ASSERT_EQ(empty, convolution<2>(empty, empty));
 }
 
+TEST(ConvolutionTest, Conv257) {
+    const int MOD = 257;
+    std::vector<ll> a(128), b(129);
+    for (int i = 0; i < 128; i++) {
+        a[i] = randint(0, MOD - 1);
+    }
+    for (int i = 0; i < 129; i++) {
+        b[i] = randint(0, MOD - 1);
+    }
+
+    ASSERT_EQ(conv_naive<MOD>(a, b), convolution<MOD>(a, b));
+}
+
 TEST(ConvolutionTest, Conv2147483647) {
     const int MOD = 2147483647;
     using mint = static_modint<MOD>;